+++ packages/react/src/components/OneFilterPicker/components/FiltersControls.tsx
@@ -1,17 +1,17 @@
 import { useEffect, useId, useState } from "react"
-import { Button } from "../../../../components/Actions/Button"
-import { Filter } from "../../../../icons/app"
-import { useI18n } from "../../../../lib/providers/i18n"
-import { Popover, PopoverContent, PopoverTrigger } from "../../../../ui/popover"
-import { getFilterType } from "../FilterTypes"
-import { FilterTypeContext, FilterTypeSchema } from "../FilterTypes/types"
+import { Filter } from "../../../icons/app"
+import { useI18n } from "../../../lib/providers/i18n"
+import { Popover, PopoverContent, PopoverTrigger } from "../../../ui/popover"
+import { Button } from "../../Actions/Button"
+import { getFilterType } from "../filterTypes"
+import { FilterTypeContext, FilterTypeSchema } from "../filterTypes/types"
 import type { FiltersDefinition, FiltersState } from "../types"
 import { FilterContent } from "./FilterContent"
 import { FilterList } from "./FilterList"
 
 interface FiltersControlsProps<Filters extends FiltersDefinition> {
-  schema: Filters
-  filters: FiltersState<Filters>
+  filters: Filters
+  value: FiltersState<Filters>
   onChange: (value: FiltersState<Filters>) => void
   isOpen?: boolean
   onOpenChange?: (open: boolean) => void
@@ -19,8 +19,8 @@ interface FiltersControlsProps<Filters extends FiltersDefinition> {
 }
 
 export function FiltersControls<Filters extends FiltersDefinition>({
-  schema,
   filters,
+  value,
   onChange,
   isOpen: controlledIsOpen,
   onOpenChange: controlledOnOpenChange,
@@ -35,10 +35,10 @@ export function FiltersControls<Filters extends FiltersDefinition>({
   const isOpen = controlledIsOpen ?? internalIsOpen
   const onOpenChange = controlledOnOpenChange ?? setInternalIsOpen
 
-  const [localFiltersValue, setLocalFiltersValue] = useState(filters)
+  const [localFiltersValue, setLocalFiltersValue] = useState(value)
   useEffect(() => {
-    setLocalFiltersValue(filters)
-  }, [filters])
+    setLocalFiltersValue(value)
+  }, [value])
 
   const updateFilterValue = (key: keyof Filters, newValue: unknown): void => {
     setLocalFiltersValue((prev) => ({
@@ -56,12 +56,12 @@ export function FiltersControls<Filters extends FiltersDefinition>({
     const getFirstFilterNotEmpty = () => {
       return Object.entries(localFiltersValue).find(([key, value]) => {
         // TODO: Make this type better
-        const filterType = getFilterType(schema[key].type) as unknown as {
+        const filterType = getFilterType(filters[key].type) as unknown as {
           isEmpty: (value: unknown, context: FilterTypeContext) => boolean
         }
 
         return !filterType.isEmpty(value, {
-          schema: schema[key] as unknown as FilterTypeSchema,
+          schema: filters[key] as unknown as FilterTypeSchema,
         })
       })
     }
@@ -98,7 +98,7 @@ export function FiltersControls<Filters extends FiltersDefinition>({
           <div className="flex h-[min(448px,80vh)] flex-col">
             <div className="flex min-h-0 flex-1">
               <FilterList
-                definition={schema}
+                definition={filters}
                 tempFilters={localFiltersValue}
                 selectedFilterKey={selectedFilterKey}
                 onFilterSelect={(key: keyof Filters) =>
@@ -108,7 +108,7 @@ export function FiltersControls<Filters extends FiltersDefinition>({
               {selectedFilterKey && (
                 <FilterContent
                   selectedFilterKey={selectedFilterKey}
-                  definition={schema}
+                  definition={filters}
                   tempFilters={localFiltersValue}
                   onFilterChange={updateFilterValue}
                 />