+++ packages/react/src/experimental/OneDataCollection/exports.ts
@@ -1,5 +1,5 @@
+export type * from "../../components/OneFilterPicker/exports"
 export type * from "./actions"
-export type * from "./Filters/exports"
 export { OneDataCollection, useDataSource } from "./index"
 export type * from "./item-actions"
 export type * from "./navigationFilters/types"